requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
google-generativeai>=0.3.2
python-multipart>=0.0.6
Pillow>=10.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
mcp>=1.0.0
mirascope[google]>=1.0.0
asyncpg>=0.30.0
asgiref>=3.9.1
unidecode>=1.4.0
